requirements.txt
Django>=5.1,<5.2
pillow
cloudinary
python-decouple
django-htmx
django-tailwind[reload]
django-markdownx
markdown
sib-api-v3-sdk
requests